client: throw non-structured response errors as internal errors
Some servers return errors in a non-v1.1+ compliant manner, as text.
An example in the wild is electrs (https://github.com/romanz/electrs).

This behaviour was apparently common in pre-v1.1 servers, and since it
is easy to support, throw the error with its text so the caller at least
gets some error context rather than "invalid server response".